Fix modeling errors on render-knowledge-graph
