+++ packages/voictents-and-estinants-engine/src/custom/programs/render-knowledge-graph/dependency/fileDependency.ts
@@ -1,3 +1,5 @@
+import { InMemoryOdeshin2IndexByName } from '../../../../core/engine/inMemoryOdeshinVoictent2';
+import { InMemoryVoque } from '../../../../core/engine/inMemoryVoque';
 import { assertNotUndefined } from '../../../../utilities/assertNotUndefined';
 import { buildNamedConstructorFunction } from '../../../../utilities/constructor-function/namedConstructorFunctionBuilder';
 import { isNotNull } from '../../../../utilities/isNotNull';
@@ -156,3 +158,21 @@ export const { FileDependencyInstance } = buildNamedConstructorFunction({
     },
   })
   .assemble();
+
+export const FILE_DEPENDENCY_GEPP = 'file-dependency';
+
+type FileDependencyGepp = typeof FILE_DEPENDENCY_GEPP;
+
+export type FileDependencyVoictentPelie = {
+  importedFileListByImportingFilePath: Map<string, BoundedFile[]>;
+  importingFileListByImportedFilePath: Map<string, BoundedFile[]>;
+  list: FileDependency[];
+};
+
+export type FileDependencyVoque = InMemoryVoque<
+  FileDependencyGepp,
+  FileDependency,
+  FileDependency,
+  InMemoryOdeshin2IndexByName,
+  FileDependencyVoictentPelie
+>;
